net: sched: Support larger rate limits
Currently the userspace "tc" program uses 32bit unsigned integers to
specify rate limits for traffic classes in bytes/second. These rate
limits are used to configure various qdiscs like HTB, TBF, HFSC etc. The
32bit rate limit used in qdiscs restricts the maximum rate limit that
can be configured to around 4GiBps = 32gibit. For larger rate limits,
"tc" silently fails by truncating the rate to 32bits and configures
incorrect rate limits.

This patch allows rate limits to take unsigned 64bit integer values
enabling use on faster interfaces such as 40gbit Ethernet. It also
requires updates to the "tc" userspace program to configure rate limits
using 64bit values.

Example with HTB qdisc:

tc qdisc add dev eth1 root handle 1: htb default 1

tc class add dev eth1 classid 1:1 parent 1: htb rate 35000Mbit \
        mtu 65535

tc class show dev eth1

With current 32bit rate limits:
class htb 1:1 root prio 0 rate 640262Kbit ceil 640262Kbit burst 65386b \
        cburst 65386b

With 64bit rate limits and updated "tc" binary:
class htb 1:1 root prio 0 rate 35000Mbit ceil 35000Mbit burst 56875b \
        cburst 56875b

+++ include/net/sch_generic.h
@@ -690,9 +690,9 @@ static inline u64 psched_l2t_ns(const struct psched_ratecfg *r,
 	return ((u64)len * r->mult) >> r->shift;
 }
 
-extern void psched_ratecfg_precompute(struct psched_ratecfg *r, u32 rate);
+extern void psched_ratecfg_precompute(struct psched_ratecfg *r, u64 rate);
 
-static inline u32 psched_ratecfg_getrate(const struct psched_ratecfg *r)
+static inline u64 psched_ratecfg_getrate(const struct psched_ratecfg *r)
 {
 	return r->rate_bps >> 3;
 }
